requirements.txt
Django
django-tastypie
iso8601plus
validictory